// types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      llm_models: {
        Row: {
          created_datetime_utc: string
          id: number
          is_temperature_supported: boolean
          llm_provider_id: number
          name: string
          provider_model_id: string
        }
        Insert: {
          created_datetime_utc?: string
          id?: number
          is_temperature_supported?: boolean
          llm_provider_id: number
          name: string
          provider_model_id: string
        }
        Update: {
          created_datetime_utc?: string
          id?: number
          is_temperature_supported?: boolean
          llm_provider_id?: number
          name?: string
          provider_model_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "llm_models_llm_provider_id_fkey"
            columns: ["llm_provider_id"]
            isOneToOne: false
            referencedRelation: "llm_providers"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      llm_providers: {
        Row: {
          created_datetime_utc: string
          id: number
          name: string
        }
        Insert: {
          created_datetime_utc?: string
          id?: number
          name: string
        }
        Update: {
          created_datetime_utc?: string
          id?: number
          name?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}